requirements.txt
black~=24.10.0
celery~=5.4.0
crispy-bootstrap5~=2023.10
Django==4.2.7
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
environs~=9.5.0
gunicorn~=21.2.0
num2words~=0.5.13
psycopg2-binary~=2.9.9
pymorphy3~=2.0.2
pymorphy3-dicts-uk~=2.4.1.1.1663094765
python-docx~=1.1.2
python-slugify~=8.0.4
redis~=5.2.0
reportlab==4.0.8
